moved components and screen to better follow folder design philosophy
